Debounce within playground so initial state restoration works
